testserver: pass options through bytes() before using with binary data.
Previously the arguments to GenerateCertKeyAndOCSP were used in operations with
binary data directly. If they were unicode objects then their encoding may not
have supported the DER data causing an error.

By passing through bytes(), we get a str object and we make it clear that we'll
be using this with binary data.

BUG=119642
TEST=net_unittests

Review URL: http://codereview.chromium.org/10188005