Split out binary file from the content shell CL.
Necessary to make it pass the try-bots, which don't support binary files.

When generating an android APK, it uses all the files in the res/ directory
and generates a R.java which contains resource IDs to be referenced in the
code, so a missing file will result in a missing ID reference and non-compiling
code.

BUG=118591
TEST=

Review URL: https://chromiumcodereview.appspot.com/10216003