Fix memory errors in sync unit tests
The commit r133754 introduced some errors.  That change made the tests
delete the session a bit more frequently, while some of the tests
maintained more pointers to its internal structures.

This change fixes all the affected tests and adds a member function to
the test framework class to try to prevent this from happening in the
future.

BUG=122033,123270
TEST=

Review URL: https://chromiumcodereview.appspot.com/10212008