Temp patch to fix offical build due to 2x.pak reference
My Metro/HiDPI change (r133613) broke the official build. The problem was that I was referencing the new 2x pak files but they aren't created by default yet.

mmoss is working on adding support for optional files. Until that is in I'm just removing the reference to the 2x pak files.

This means that for now 2x pak files won't be archived but that's not a big deal since we don't build with HIDPI_ENABLED anyways.

BUG=114311
TEST=

Review URL: https://chromiumcodereview.appspot.com/10207024