Change NaClBrokerService to use weak pointers to NaClProcessHost
This ensures that if the NaCl process crashes before
NaClBrokerService::OnDebugExceptionHandlerLaunched() is called, this
method does not use a dangling NaClProcessHost pointer.

This is in preparation for changing NaCl's Windows debug exception
handler to attach on demand, which would allow untrusted NaCl code to
cause the NaCl process to exit before
OnDebugExceptionHandlerLaunched() runs.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2618
TEST=nacl_integration etc.

Review URL: https://chromiumcodereview.appspot.com/10227013