Cope with the hostId being unknown in connectMe2Me().
This improves the user experience when a tab is reloaded containing a Chromoting session, and tries to re-connect to a host that has been unregistered in the meantime.  Previously the hostId lookup would fail and the Chromoting UI hang due to an unhandled JavaScript exception.

Also updated a comment in connectMe2MePin() for a similar validity check, to clarify the circumstances in which it may fail.

BUG=123898


Review URL: http://codereview.chromium.org/10246006