gdata: Add GDataFileSystem::GetEntryByPathAsync()
In preparation for making FindEntryByPathAsync() private.
The new function is not used in production code yet.

This function is needed as we get metadata of a file or
a directory in gdata_file_system_proxy.cc, without knowing
if the file path points to a file or a directory.

BUG=chromium-os:30251
TEST=add unit tests

Review URL: https://chromiumcodereview.appspot.com/10274025