ppapi Graphics2d should invalidate all changes without clipping to visible area.
Currently the flush operation assumes that WebKit will request the ppapi plugin
to repaint any newly exposed regions, and so it only invalidates the visible
portion of changes. However, for accelerated compositing, this behavior is not
correct. This patch rearranges the logic to work for both accelerated and
non-accelerated compositing.

BUG=122860
TEST=tested manually, created crbug.com/125763 for making a unit test


Review URL: http://codereview.chromium.org/10223009