Skip InitializeSandbox for out-of-sandbox plugins
On ChromeOS R20, the gtalk plugins failed to load after some changes to the
setup of seccomp filter sandbox. We now append the --no-sandbox flags for
PPAPI plugins that need to run out of the sandbox, so the PPAPI process does not
apply sandboxing.

BUG=chromium-os-partner:9390
TEST=Run gtalk video on ChromeOS


Review URL: http://codereview.chromium.org/10295002