Make URLRequestThrottlerManager a member of URLRequestContext.
This addresses a long-standing TODO and allows us to simplify the
logic in the class a bit as it now lives fully on the IO thread.  It
should also allow further cleanup in follow-up changes e.g. to stop
using scoped_refptr for the URLRequestThrottlerEntry instances.

BUG=119760


Review URL: http://codereview.chromium.org/10203002