Conflict detection for Extension Keybinding.
At install time, the extension will now try to make its suggested keybindings active. If the keybinding is free, then the extension is successful (before this CL the last registered extension would win). If the keybinding is not free, the keybinding becomes inactive. I've introduced a pref that keeps track of which keybinding is active, which will become the basis for a config UI.

This changelist also fixes a DCHECK on close of Chrome if an extension with a keybinding has been uninstalled (we were unregistering twice).

BUG=121419
TEST=Covered in automated test. (Install two extensions with the same keybinding. Only the first should get its keybinding activated).
Review URL: https://chromiumcodereview.appspot.com/10201016