NTP: Add a drop-down menu to the Other Devices list.
You can currently right-click on session names to bring up a context
menu. This is very undiscoverable, so we want drop-down buttons beside
the session names that will show the menu when clicked.

This functionality is that same as on chrome://history, so pull the code
out of history and into shared JS & CSS files.

BUG=124824
TEST=Manual. Open the Other Devices menu, click on one of the drop downs,
and ensure that the Hide menu item works properly.

Review URL: https://chromiumcodereview.appspot.com/10327002