Deleting NOTREACHED() in ExtensionFunction::HandleBadMessage().
There does not seem to be a reason to have it there.
Handling the bad message should be the same in Release and Debug modes,
having the NOTREACHED() would cause:
 * browser crash in Debug
 * extension process crash in Release

History of that particular NOTREACHED():
 * appeared in ExtensionFunction::HandleBadMessage() in
   http://codereview.chromium.org/6927076, moved from
   ExtensionFunctionDispatcher::HandleBadMessage()
 * was introduced as
   ExtensionFunctionDispatcher::HandleBadMessage()
   in http://codereview.chromium.org/77003

I had not found any particular reason in the above CLs why
we would like to have the above difference in behaviour.

BUG=115467
TEST=N/A


Review URL: http://codereview.chromium.org/10358002