Omnibox: Add logging of whether the user just deleted text from the omnibox
For UMA opted-in users, this adds a new field to the omnibox event record that's uploaded when the user uses the omnibox.  In particular, it records a simple boolean of whether the user just deleted text in the omnibox before accepting a suggestion.  Basically, was the last key in the omnibox a backspace?

This change only adds the field to the protocol buffer upload format, not the XML format. If I decide later I want to add it to the XML format, I'll do it in another changelist.

BUG=
TEST=by hand, with VLOG(1) lines in autocomplete_edit.cc


Review URL: http://codereview.chromium.org/10337005