Fix getLastFinishedBuild() so it works in multi-slave per builder cases.
Fix getLastFinishedBuild() so it works correctly when there are 2 or more
running builds for a given builder.  Previously, it would always return the
second-to-last build if the first was missing or running.  Now it always
returns the build that is just after what we assert is the oldest running
build.

BUG=126251
R=maruel@chromium.org,szager@chromium.org


Review URL: http://codereview.chromium.org/10365024