Fix Instant pyauto tests.
http://crrev.com/135298 makes --instant-field-trial override everything else,
including the kInstantEnabled pref. Since test_launcher_utils.cc adds the
switch (with a 'disabled' value) to all tests, we have to override the switch
when trying to enable Instant.

The same CL also made preloading of the search engine the default. So, even if
you type a non-search string (such as a URL or "javascript:") into the omnibox,
Instant's preview (overlay) will still be active. To properly test that Instant
isn't actually showing the preview, test for 'current' instead of 'active'.

BUG=126261
TEST=instant.py


Review URL: http://codereview.chromium.org/10379014