Enable caching and using cached json data in console.
One last item was not being cached by chromium-build:
JSON data from chromium.lkgr which clients load at
page-render-time.  This CL adds support to chromium-build
to fetch this data and to reference it from its local
store so the client retrieves it from chromium-build.

Other changes:
  - Make chromium-build references in console https by default.
  - Fix line lengths so they're under 80 chars per line.
  - Remove unused master argument from ConsoleData class.
  - Disregard unused variable checks for handlers.
  - Update test data to match new expectations based on these
    changes.

With the https changes in this CL, viewing
https://chromium-build.appspot.com/p/chromium/console results
in a green lock in Chrome since all of its included resources
are also fetched over SSL/TLS.  It used to result in a broken
lock.

TBR=maruel@chromium.org


Review URL: http://codereview.chromium.org/10386004