Added a SetForced method to allow forcing a new set of group bucketting.
This will be needed by the Finch client (soon to be known as the Variations Service) that will get updated information about FieldTrials and pre-feed the field trial list based on this information.

So this is different from CreateFieldTrial, which forced a group choice, this new method allows forcing a new set of appended groups.

BUG=121695
TEST=build\Debug\base_unittests.exe --gtest_filter=FieldTrialTest.SetForced


Review URL: https://chromiumcodereview.appspot.com/10382018