Upstream youtube plugin for chrome on android
For android devices, flash plugin is no longer supported.
To support oldstyle youtube embeds for chrome on android, we display
a poster and link the it to the mobile youtube app.
So if the plugin is not found, we check if it is a youtube link and
bring up the youtube app.

BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/10273020