Make GoogleURLTracker a profile-keyed service.
The original multi-profile work made this object use the system request context, which meant it never sent cookies to searchdomaincheck.  This is bad because these are important to determining the correct domain to reply with.

Instead, there should be a different instance per-profile, and these instances should use the standard Profile-scoped cookie stores, so we send the relevant google.com cookies.

BUG=106086
TEST=Requests to searchdomaincheck should send cookies.
Review URL: https://chromiumcodereview.appspot.com/10377054