Reland "Implement random wallpaper feature"
CL 10302016 reverted because of failed testes on linux chromeos bot. The reason
was a DCHECK fail. I have fixed the fails in CL 10382072. So reland the random
wallpaper feature here.


TBR=nikita,ben,jhawkins
BUG=120217
TEST=Go to set wallpaper overlay in settings page, check the "I'm feeling
lucky" checkbox. Sign out and sign in a couple of times, verify if the
wallpaper is random.


Review URL: https://chromiumcodereview.appspot.com/10384079