Remove node_leak_test.cc from test_shell_tests.
This is not an automated test: if you pass --test-url to test_shell_tests,
it tries to load the URL then exits.  The comment makes it sound like this
was for manual testing to check for node leaks (seeing the node count
printed to stdout), but there's no need for this to be in test_shell_tests.

I doubt this is being used anymore (you can't load http because it's
blocked and we no longer output the node count), so just deleting it.

BUG=126514

Review URL: https://chromiumcodereview.appspot.com/10375060