Revert 136124 - Allow the RWHVW to be focused when a window is restored from minimized.
This restores a hack that was in place in the old NativeTabContentsContainer code that seemed to clear focus during restoration (though the hack was for another purpose, the hack seemed to fix this bug too). I've isolated the hack to the re-activation code.

http://crbug.com/125976
TEST=see bug
Review URL: https://chromiumcodereview.appspot.com/10378043

TBR=ben@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10377070