Fix dropped Play event when there's a pending pause.
If the message queue is built in such a way that DoPlay()->DoPause()->DoPlay()
ends up executing before the the first DoPlay()'s PollAndStartIfDataReady() runs,
we'll end up dropping the play and keeping the pause.

Fixed by ensuring we drop the pause if we receive a play before the pause is
consumed by PollAndStartIfDataReady().

BUG=111272
TEST=unittests, test.html test case in bug. Manual seeking in buffered areas.

Review URL: https://chromiumcodereview.appspot.com/10386037