Mac Pepper Flash: C++ wrapper for VideoDecodeAcceleration.framework
This CL adds a C++ wrapper for the VideoDecodeAcceleration framwork. This will be used by the pepper flash port for hardware accelerated video decoding.

I'm adding this to ui/gfx instead of content/common/gpu/media because I think it might be useful for othercases as well.

BUG=127414
TEST=Verified that test passed on my MacPro (without hardware support) and MacBook Pro (with hardware support).

Review URL: https://chromiumcodereview.appspot.com/10152009