NaCl: Fix ppapi_geturl_browser_test to run using the IRT
This test was using an old-style library list so it was running
against a version of libppruntime supplied by the web app instead of
the version built into the IRT library supplied by the system.

This change should save a lot of build time in nacl_integration, which
should no longer compile libppruntime.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2648
TEST=nacl_integration

Review URL: https://chromiumcodereview.appspot.com/10383096