Remove the 'Test Finished' step from run_tests.py.
The 'Test Finished' step was introduced in r130823 to absorb the
non-zero exit status which we'd get if any of the test suites in
the batch would fail.
Since the run_tests.py script will print a STEP_FAILURE annotation
for every suite the exit status is redundant and therefore should
be removed.

BUG=none
TEST=none


Review URL: https://chromiumcodereview.appspot.com/10382100