dbus: revamp fd passing support for i/o restrictions
Encapsulate file descriptor validity checking and status in the companion FileDescriptor class so callers can do descriptor checking in a context where i/o is allowed.
    
Update the debug daemon client support to validate the pipe descriptors in a worker thread so it is not done on the UI thread.

BUG=126142

TEST=new unit tests + collect trace data on chrome os and verify no assert is triggered

Review URL: https://chromiumcodereview.appspot.com/10382021