The current century—spanning the years from 2001 to 2100—referred to as the 21st century. The atomic number of scandium is 21 (3 more to go for Chromium) The standard TCP/IP port number for FTP connection
In most USA states 21 is the drinking age.
The Twenty-first Amendment repealed the Eighteenth Amendment, thereby ending Prohibition.
A card game, also called vingt-et-un (French for "twenty-one"), pontoon, or blackjack
The number of spots on a standard cubical (six-sided) die (1+2+3+4+5+6), not so useful for blackjack, but perfect for craps
21 is the sum of the first six natural numbers (1+2+3+4+5+6=21).
Twenty-one is a Fibonacci number, a Harshad number, a Motzkin number, a triangular number and an octagonal number, as well as a composite number, its proper divisors being 1, 3 and 7.

Number 21 is the name of the plane alleged flown by Gustave Whitehead two years before the Wright brothers' flight

Credit: http://en.wikipedia.org/wiki/21_(number)