Add scripts to list or trace all test cases in a gtest executable.
list_test_cases.py quickly outputs the test cases in a gtest executable. For browser_tests is takes ~100ms.

trace_test_cases.py traces each test individually to be able to generate pretty graphs.

read_trace.py is a debugging tool to read a single trace.

Refactor trace_inputs.py to make it possible. I changed the semantic for it so that stdin is closed and stderr is redirected to stdout. I also have the logger not look at the logging level to determine if output should be shown.

R=rogerta@chromium.org
BUG=98636
TEST=

Review URL: https://chromiumcodereview.appspot.com/10377105