If a navigation was initiated by the browser, store the referrer policy to use in the document state
We can't put it into the document, because if the navigation was cancelled, we had to restore the original referrer policy.

Also, when trying to get the referrer policy for a given request, fall back to the document's referrer policy if the request wasn't yet tagged by willSendRequest. e.g. in loadURLExternally() the request didn't pass through willSendRequest

BUG=124750
TEST=Enabled the previously failing tests in ReferrerPolicyTest.*

Review URL: https://chromiumcodereview.appspot.com/10392041