[Sync] Replace hardcoded gaia urls in sync_test.cc with calls to GaiaUrls
In the sync integration tests, we stub out calls to various gaia urls so
that the tests are self contained, and don't have a dependency on
external services.

The hardcoded urls in sync_test.cc are now available for use via a
class called GaiaUrls. This patch gets rid of the hardcoded urls.

BUG=none
TEST=sync_integration_tests


Review URL: https://chromiumcodereview.appspot.com/10387069