[Chromoting] Fixup the me2me_host plist so that the application is valid.
Fix the plist file by removing the '@'s from the version strings.

This is required for the .app bundle to be recognized as a valid app by the OS.
Without this fix, the official signing fails and the app icon is displayed as
the "broken app" icon.

In addition, this adds a default icon and sets the human readable copyright
message. And moves the COPYRIGHT_BY variable to the top of the .gyp so that
it can be shared by multiple targets.
Review URL: https://chromiumcodereview.appspot.com/10381113