Verify that xmpp_login specified in the config matches auth token.
Previously the host would ignore the case when the oauth token doesn't
match the xmpp_login specified in the config. That would lead to situations
when host is connected with one account, but verifies incoming connections
with a different account. Fix this by verifying that the JID that we receive 
from XMPP server matches the value in the config.


Review URL: https://chromiumcodereview.appspot.com/10378110