Implement IBusText
The IBusText is corresponding to text object used in dbus communication with
ibus-daemon.

According to this CL, IBusText will be compiled and tested by
chromeos_unittests, but not used in production binary at this moment.

This CL is dependent to http://codereview.chromium.org/10332022/

TEST=chromeos_unittests, unit_tests, dbus_unittests
BUG=chromium-os:26334


Review URL: https://chromiumcodereview.appspot.com/10381025