Remove VIEW_TYPE_WEB_CONTENTS and make default view type INVALID.
The view type enum is supposed to represent the 'type of view' in the user
sense, so 'tab', 'background page', 'popup', etc. A 'web contents' is not
a type of view, it's an implementation detail of many different view
types.

This was essentially resulting in too many WebContents instances being
treated as tabs.

Review URL: https://chromiumcodereview.appspot.com/10387110