Merge 136466 - chromeos: Fix user whiltelist  not properly populated regression.
This happens because the whitelist pref value is stored in policy and policy set
on device is asynchronous. Before previous policy write comes back, pref value
cache is stale and should not be used as a basis for further mutating (as what
happens when we AppendToList while there is a previously queued AppendToList).

BUG=127215
TEST=Verify fix for issue 127215.

Review URL: https://chromiumcodereview.appspot.com/10332094

TBR=ivankr@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10398021