Fix RenderText display rect y-axis adjustment in print preview
RenderText is displayed inside its display rect. Display rect's height should be used to calculate whether the display rect overlap with content canvas, and it should be used to do the adjustment as well. The checking and adjustment should not be related with font size.

Note: due to the sandbox and NativeTheme issues in Linux, I did not enable it here.

BUG=125661
TEST=manual try print preview in Windows and Linux (with enabling), all 4 components in footer/header should be rendered.

Review URL: https://chromiumcodereview.appspot.com/10392061