Merge 135934 - Preference Pane for chromoting on Mac
This adds a new System Preference pane, which is shown when a user configures
the Chromoting Host via the web-app.

This is basically functional, but is not yet properly integrated with the
flow of the web-app.  I will prepare a separate CL that implements
notifications from the pref-pane back to the web-app.

BUG=120903,121749
TEST=Manual

Review URL: https://chromiumcodereview.appspot.com/10171020

TBR=lambroslambrou@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10398029