aura/ash split: Moves RootWindowEventFilter and InputMethodEventFilter to ui/aura/.
All of InputMethodEventFilter and more than half of
RootWindowEventFilter were copied verbatim to a
DesktopRootWindowEventFilter. Instead of copying the other half for
another patch, I'm just moving RWEF and deleting the desktop version.

(InputMethodEventFilterUnittest was simple enough to move with the
file, but RootWindowEventFilter seems to act more like an integraiton
test of a lot of the stuff on Shell.)

BUG=125106

Review URL: https://chromiumcodereview.appspot.com/10332145