Merge 135921 - Remove the TemplateURL copy constructor and assignment operator.
It would be nice to not have to add CopyFrom() as a substitute but that would require completely rearchitecting TemplateURLService to never update TemplateURLs in place.  That's actually potentially feasible but I'm not going to try to tackle it.

BUG=none
TEST=none
Review URL: https://chromiumcodereview.appspot.com/10383072

TBR=pkasting@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10386153