Use third_party/jsoncpp in Mac pref-pane
Use third_party JSON library instead of base/json to remove a dependency on
code from "base".  This is to enable the pref-pane to be built for 64-bit on
Mac OS X.

Linking to individual files in base/ is not feasible, since any calls to the 
logging facility cause many files to be pulled in as dependencies, creating a 
burden for any maintainers of those files. So we prefer to avoid using base/
entirely, until Chrome supports building for 64-bit on Mac OS X. 


BUG=125116
TEST=manual


Review URL: https://chromiumcodereview.appspot.com/10381141