Move declarative API into events API
The motivation of this CL is to bring the declarative API and the rest of the
events mechanism closer together in the API documentation. To achive this, we
move the content of declarative.json and everything related to events from
extension.json into the events.json API definition. The new "events" namespace
is automatically provided to all extensions.

This is a second attempt after
http://src.chromium.org/viewvc/chrome?view=rev&revision=137388 failed in
ExecuteScriptApiTest.ExecuteScriptInFrame. TBR'ing as this is a tiny change from the previous CL.

BUG=126497,112155
TEST=no
TBR=aa@chromium.org, kalman@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10392127