Add a autotest dependency for Page Cycler data
This creates a new dependency which will be used by ChromeOS
to package up the latest set of Page Cycler data from the
Chrome source tree.  The Chrome ebuild does the heavy lifting
of getting this packaged up, but this provides the script
which will get called on the target that puts a symlink in
the correct place to the pc data files.

BUG=chromium-os:23990
TEST=none

R=nirnimesh@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10381167