Add PyAuto test for DeviceGuestModeEnabled policy
This CL adds a PyAuto test for the DeviceGuestModeEnabled policy which
enables/disables guest login on Chrome OS devices.
As a prerequisite, the browser automation framework is fixed so that it
does not wait for the login screen to be shown when the account picker
is enabled instead.

BUG=128343
TEST=New PyAuto ChromeosDevicePolicy.testGuestModeEnabled


Review URL: https://chromiumcodereview.appspot.com/10382197