NaCl: Re-enable nacl-glibc tests on Windows (retry)
These tests were disabled before because libppruntime.so was failing
to link because the link command sometimes hit a command line length
limit.  We have since worked around the length limit, and
libppruntime.so should no longer be built by the tests anyway.

This was previously committed as r136913 but it failed on the
buildbots and was reverted because of a compile problem with
__cpuidex() that has since been fixed.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2648
TEST=nacl_integration

Review URL: https://chromiumcodereview.appspot.com/10389182