Fix logging of string16s from gfx namespace on Windows.
I don't why this fixes it, but somehow having a SelectionModel << logging operator in the gfx namespace makes Visual Studio fail to find the << logging operator for string16 arguments.

Per discussion with Scott, we decided to remove the << operators altogether and replace them with a ToString() function to be consistent with Rect, Size, etc.

TEST=Check that adding a "LOG(INFO) << text();" line to RenderTextWin compiles.

Review URL: https://chromiumcodereview.appspot.com/10332205