Compute the enclosing rectangle for circles/ellipses generated from a series of touch events
For a sequence of {touchstart, touchmove*, touchend} events, compute the rectangle that encloses all of the circles/ellipses corresponding to the touch events. If the hardware is capable of collecting radius_x or radius_y, use the maximum of these two values as the radii values for the touch region. Otherwise use a default radius value of 15.

BUG=122190
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10262023