Ensure that the at-launch reauthorize actually reauthorizes all items, even if an at-update reauthorize already ran.
Although the at-update reauthorization renders items readable by the
application when signed by either certificate, the icon and application name
displayed in Keychain Access are wrong, because they refer to the stub
reauthorization executable on the ejected updater dmg. There is no functional
problem here, but as long as it remains possible to reauthorize at launch,
it's good to do so to fix the UI in Keychain Access.

BUG=108238
TEST=none
Review URL: https://chromiumcodereview.appspot.com/10388184