Turn on perf graphs for buildbot test named pyauto_perf_tests, and remove ChromiumQAFactory.
I have another pending CL that is internal to Chrome which removes all
uses of ChromiumQAFactory.  It is removed in this CL since it is no longer
used.

BUG=chromium-os:22239
TEST=Verified the pyauto_perf_tests step will output perf data and put a
"results" link on the waterfall test step with this change, using a local
master/slave buildbot configuration.
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/10386193