Add deletion observer to P2PSocketDispatcher.
This is an alternative solution to http://codereview.chromium.org/10332114/

RenderViewObserver is actually destroyed in the same order as they are created. 
Ie 
P2PSocketDispatcher::P2PSocketDispatcher 
MediaStreamImpl::MediaStreamImpl 
P2PSocketDispatcher::~P2PSocketDispatcher 
MediaStreamImpl::~MediaStreamImpl 

Since MediaStreamImpl holds an observer object to P2PSocketDispatcher and the P2PSocketDispatcher require all observers to have unregistered before it is deleted, it is required that the MediaStreamImpl is either destroyed before P2PSocketDispatcher or that it is notified that the 
PSocketDispatcher will be destroyed. 

BUG=127782
TEST= Please see the defect


Review URL: https://chromiumcodereview.appspot.com/10387135