aura: Fix the sythentic mouse-event location.
The synthetic mouse event is generated at the last mouse DIP-location. But this
event gets transformed for DIP again before it gets dispatched. This ends up
sending the event with incorrect location. So make sure it starts off with the
pixel-location so that the event location gets transformed for DIP only once.

BUG=128095
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10408037