Merge 138057 - gdata: Add requestDirectoryRefresh to file_browser_private.
This is a reland of r138055 with a fix for DCHECK failures.

The new API is used to fetch the metadata of files in a particular
directory. More specifically, the new API is used to fix the stale
thumbnail URL issue.

BUG=127697
TEST=added a unit test.

Review URL: https://chromiumcodereview.appspot.com/10386226

TBR=satorux@chromium.org

NOTE: fileBrowserPrivate.json was manually edited as the file was renamed to
file_browser_private.json on trunk