Revert 138166 - Merge 137824 - Properly handle accounts that don't have GMail account.
1. Me2MeHostAuthenticatorFactory now verifies that the bare 
JID of the remote client matches bare JID of the host. 
Previously it was comparing it with the user's email, 
which may be different from JID.
2. GaiaOAuthClient now fetches user's email.
3. SignalingConnector verifies that user's email matches 
the expected value stored in xmpp_login. If it doesn't, 
then the auth token was generated for a different account 
and the host treats it as an authentication error.

BUG=128102

Review URL: https://chromiumcodereview.appspot.com/10332187

TBR=sergeyu@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10414033

TBR=sergeyu@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10417029