Desktop aura: Put x11 cache in aura, instead of ui/base/x/.
This should solve the pyauto compile error because MessagePumpX is actually an aura component, not a general x11 component.

Instead of being a Singleton, X11AtomCache should tie its lifetime to an aura::Env, which are reset in the unit tests whenever an X server is switched out.

BUG=128584, 129075
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10416020