Defer CGLFlushDrawable until OSX-requested drawRect to avoid spinning when window is obscured.
Sadly, this fix regresses some black flashing bugs:
 -flicker-test2.html occasionally flashes while resizing the window.

I think this regression could be fixed if we always draw with the OpenGL context once the context is created. But that would be a third drawing path and some significant code, so let's consider it in a follow up.

BUG=127709
TEST=open https://cvs.khronos.org/svn/repos/registry/trunk/public/webgl/sdk/demos/google/particles/index.html ; open Activity Monitor and observe the chrome process CPU usage ; resize chrome window so it's smaller than the Activity Monitor window (or increase size of Activity Monitor); move Activity Monitor on top of chrome window to fully obscure it ; verify that the CPU usage of chrome does not spike upwards while the chrome window is obscured.

Review URL: https://chromiumcodereview.appspot.com/10382213