Reformatting only CL; NO CODE CHANGE.
Sort the test names.
Remove commented out code in master.chromium.git.

This CL is in preparation to renaming tests to their executable name so it
starts to make sense. Doing so will cause a lot of string expansion, which would
cause a lot of list reformatting. So to simplify the following CLs, reformat
first.

R=nsylvain@chromium.org
BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10408088