Remove non-functional 'Slave to use' field.
'Slave to use' is non-functional in Buildbot 0.8.  A
property slavename will get set on a buildset, and
available to a build that's running as a property on
that build, but where it's needed is in the buildrequest
which will be used to trigger the build and determine
which builder and slave to run the build on.

R=maruel@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10426004