Remove the need for ResourceDispatcherHostTest to call PauseRequest.
It turns out that the PausedCancel test was testing an impossible case.  Since
ResourceDispatcherHostImpl::PauseRequest is a private mtehod, it is no longer
possible to pause a request at arbitrary times.  It is only possible to pause
from ResourceHandler methods that have a defer out param.  As a result, this
PausedCancel test can simply be removed.

ResourceDispatcherHostTest::ResumeRequest could just be removed since there
is no caller.

The PausedStartError was modified to use a ResourceThrottle implementation to
pause the request.

With these changes, only ResourceDispatcherHostImpl can call PauseRequest.

R=jam@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10422009